modifica clase que inicia a grabacion que la guarde como array de byte

+++ app/src/main/java/com/newventuresoftware/waveformdemo/AudioDataReceivedListener.java
@@ -15,5 +15,5 @@
 package com.newventuresoftware.waveformdemo;
 
 public interface AudioDataReceivedListener {
-    void onAudioDataReceived(short[] data);
+    void onAudioDataReceived(byte[] data);
 }
